requirements.txt
# ================================
# TRADING BOT - REQUIREMENTS
# Install: pip install -r requirements.txt
# ================================

# CORE (REQUIRED) - Install these first
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
ta>=0.11.0

# MACHINE LEARNING (Optional)
scikit-learn>=1.3.0

# VISUALIZATION (Optional)
matplotlib>=3.7.0
plotly>=5.18.0

# WEB FRAMEWORK (Optional - for dashboard)
flask>=3.0.0
flask-cors>=4.0.0

# UTILITIES (Optional)
python-dotenv>=1.0.0
colorlog>=6.8.0
requests>=2.31.0

# SENTIMENT ANALYSIS (Optional)
textblob>=0.17.1
beautifulsoup4>=4.12.2

# DATABASE (Optional)
sqlalchemy>=2.0.0

# TESTING (Optional)
pytest>=7.4.0

# ================================
# INSTALLATION NOTES:
# ================================
# 
# OPTION 1 - Install everything:
#   pip install -r requirements.txt
#
# OPTION 2 - Install only essentials:
#   pip install pandas numpy yfinance ta
#
# OPTION 3 - Install one by one:
#   pip install pandas
#   pip install numpy
#   pip install yfinance
#   pip install ta
#
# ================================
# DO NOT INSTALL:
# - TA-Lib (requires C++ compilation)
# - pandas-ta (not needed, we use 'ta')
# - tensorflow (too large, optional for ML)
# ================================